requirements.txt
gspread==5.10.0
oauth2client==4.1.3
python-dotenv==1.0.0
yagmail==0.15.293
openai>=1.0.0
